backend/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
boto3==1.34.0
django-cors-headers==4.3.1
Pillow==10.1.0
scikit-learn==1.3.2
xgboost==2.0.3
pandas==2.1.4
joblib==1.3.2
channels==4.0.0
channels-redis==4.1.0
django-ratelimit==4.1.0
python-dotenv==1.0.0